separate service API in C into multiple files.
