+++ include/dsn/cpp/task_helper.h
@@ -36,7 +36,6 @@
 # pragma once
 
 # include <dsn/service_api_c.h>
-# include <dsn/ports.h>
 # include <dsn/cpp/auto_codes.h>
 # include <dsn/cpp/utils.h>
 # include <dsn/cpp/rpc_stream.h>
