+++ src/core/core/library_utils.h
@@ -35,7 +35,7 @@
 
 # pragma once
 
-# include <dsn/ports.h>
+# include <dsn/internal/ports.h>
 # include <dsn/cpp/auto_codes.h>
 
 # ifdef __TITLE__
